Add a field tracking the term sign date
This field will be used to track the latest date when the member agree on the ccl terms. Each member is required sign the terms before school open every year. They only need to sign the terms once under their account each year. If they refuse to agree on the terms, they would not be able to register classes or attending school classes.